pothole-detection-system/backend/requirements.txt
fastapi
uvicorn
pillow
ultralytics
opencv-python
python-multipart
streamlit
requests
plotly